refactor(client-implementation): move browser to game directory
