+++ src/client/entities/common/index.ts
@@ -9,5 +9,4 @@ export * from "./player";
 export * from "./vehicle";
 export * from "./worldObject";
 export * from "./router";
-export * from "./browser";
 export * from "./camera";
