+++ src/client/game/common/browser/IBrowserManager.ts
