+++ src/client/game/common/index.ts
@@ -1,0 +1,1 @@
+export * from "./browser";
