+++ src/client/index.ts
@@ -2,3 +2,4 @@ import { RockMod } from "./RockMod";
 
 export { RockMod };
 export * from "./entities";
+export * from "./game";
